% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driveBC_map_route.R
\name{driveBC_map_route}
\alias{driveBC_map_route}
\title{Leaflet Map}
\usage{
driveBC_map_route(start = "Kelowna", dest = "Vernon")
}
\arguments{
\item{start}{City you are starting from}

\item{dest}{City you want to arrive at}
}
\value{
A data frame containing a detailed description of each road event, a label for each road event (eg. POOR_VISIBILITY), the name of the specific highway, as well as the severity level

An interactive map that can be opened in viewer and provide severity level, and road event information
}
\description{
Builds an interactive map displaying the active road conditions between starting and destination cities
}
\examples{
driveBC::driveBC_map_route(start="Kelowna",dest="Vernon")
}
