% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driveBC_map_major.R
\name{driveBC_map_major}
\alias{driveBC_map_major}
\title{Leaflet Map}
\usage{
driveBC_map_major(days, region = NA, status = "ALL")
}
\arguments{
\item{days}{Number of days to include in search, in addition to today}

\item{region}{Default is \code{NA} which returns displays all of BC.  If you want a specific region, enter a
number that corresponds to one of the following regions.
\enumerate{
\item Lower Mainland
\item Vancouver Island
\item Rocky Mountain
\item West Kootenay
\item Okanagan-Shuswap
\item Thompson-Nicola
\item Cariboo
\item Peace
\item Fort George
\item Bulkley Stikine
\item Skeena
}}

\item{status}{Incident status.  Default is \code{ALL} which returns both ACTIVE and ARCHIVED.  If only want ACTIVE or ARCHIVED, specify in function call}
}
\value{
An interactive map that can be opened in viewer and analyzed further
}
\description{
Builds an interactive map displaying the locations of active and historical major incidents
}
\examples{
driveBC_map_major (days = 10,region = 1,status = 'ARCHIVED')
driveBC_map_major (0,1,'ACTIVE')
driveBC_map_major(15)
}
